DosWord: ignore headers and footers in body of text for now
In Word for DOS, at any point a header or footer can be defined; this
means that the following page will have this header and/or footer.

In the MS Write format, only headers and footers can be defined at the
very beginning of the document. Enforce this for both formats for now;
this fixes the parser dropping large sections of text before headers
or footers.